src/system/PPDX/OfflineIndexing/offline_indexing_requirements.txt
# Core ML and NLP dependencies
sentence-transformers>=2.2.0
huggingface-hub>=0.17.0
scikit-learn>=1.3.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Database
neo4j>=5.13.0

# Utilities
python-dotenv>=1.0.0
pathlib>=1.0.0

# Development and testing
pytest>=7.4.0
